do not fail if pid file does not exist

+++ stop.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+if [ ! -f  pid ]
+then
+    exit 0
+fi
+
 PID=`cat pid`
 echo "Stopping pid $PID"
 kill -9 $PID
